% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy10.R
\docType{data}
\name{italy10}
\alias{italy10}
\title{Italian Survey on Household and Wealth, 2010}
\format{A data frame with 11 variables: \code{parent}, \code{sex}, \code{anasc}, \code{nascreg}, ' \code{cit}, \code{acom4c}, \code{studio}, code{q}, code{qual}, code{sett}, code{ireg}. There is an additional variable \code{id} that is the unique identifier for the data set.
Common models

This data set has one year's worth of data.
This data set joined with another year's worth of 
is appropriate for performing various types of record linkage
and can be assessed by standard record linkage methods.}
\usage{
italy10
}
\description{
The 2008 survey on covered 13,702 Italian Survey on Household and Wealth
covering individuals.
}
\examples{
head(italy10)
dim(italy10)
}
\references{
Steorts (2015), Bayesian Analysis
}
\keyword{datasets,}
\keyword{entity}
\keyword{linkage,}
\keyword{record}
\keyword{resolution}

